Use 'hugow' a platform independent wrapper for Hugo binary
This will remove the need to actually have the specific
binary installed on differnet machines (users, CI, etc).

It will also give maintainers control of which version of
binary is needed to be installed (by defining the required
version in `.hugo/version`). As a result it will ensure
consistency across all different machines which need to
run the hugo binary.

So upgrading or downgrading Hugo version will be just to
change one file and commit it, the next time wrapper runs
it will download the specified version for the platform
it runs on.

And since there won't be a need to have the binaries for
different platforms checked in, the repo size will be
reduced and cloning and checking out will be faster.